Delete DML and EXEC ast nodes
